# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create user",
        "description": "Create a new user account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "StrongPassword123!"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "firstName",
                  "lastName"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "List users",
        "description": "Retrieve all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "description": "Update a user account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "updated@example.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "Jane"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Smith"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "description": "Delete a user account by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/deals": {
      "post": {
        "operationId": "DealsController_create",
        "summary": "Create deal",
        "description": "Create a new deal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deal created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Deals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DealsController_findAll",
        "summary": "Get all deals",
        "description": "Get paginated list of deals",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 10,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "example": "created_at",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deals retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Deals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deals/{id}": {
      "get": {
        "operationId": "DealsController_findOne",
        "summary": "Get deal by ID",
        "description": "Get a specific deal by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deal ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deal retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Deal not found"
          }
        },
        "tags": [
          "Deals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "DealsController_update",
        "summary": "Update deal",
        "description": "Update an existing deal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deal ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deal updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Deal not found"
          }
        },
        "tags": [
          "Deals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DealsController_remove",
        "summary": "Delete deal",
        "description": "Delete a deal by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Deal ID",
            "schema": {
              "example": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deal deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Deal not found"
          }
        },
        "tags": [
          "Deals"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/dashboard": {
      "get": {
        "operationId": "DashboardController_getStats",
        "summary": "Get dashboard stats",
        "description": "Retrieve aggregated dashboard statistics for the authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Dashboard"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "CRM Backend API",
    "description": "API documentation for the CRM backend application. This API provides endpoints for managing companies, contacts, deals, and user authentication.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDealDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Deal title",
            "example": "Enterprise Software License"
          },
          "description": {
            "type": "string",
            "description": "Deal description",
            "example": "50-seat license for enterprise CRM software"
          },
          "value": {
            "type": "number",
            "description": "Deal value in dollars",
            "example": 50000
          },
          "stage": {
            "type": "string",
            "description": "Deal stage",
            "example": "negotiation",
            "enum": [
              "lead",
              "qualified",
              "proposal",
              "negotiation",
              "closed-won",
              "closed-lost"
            ],
            "default": "lead"
          },
          "companyId": {
            "type": "number",
            "description": "Associated company ID",
            "example": 1
          },
          "contactIds": {
            "description": "Array of associated contact IDs",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateDealDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Deal title",
            "example": "Enterprise Software License"
          },
          "description": {
            "type": "string",
            "description": "Deal description",
            "example": "50-seat license for enterprise CRM software"
          },
          "value": {
            "type": "number",
            "description": "Deal value in dollars",
            "example": 50000
          },
          "stage": {
            "type": "string",
            "description": "Deal stage",
            "example": "negotiation",
            "enum": [
              "lead",
              "qualified",
              "proposal",
              "negotiation",
              "closed-won",
              "closed-lost"
            ],
            "default": "lead"
          },
          "companyId": {
            "type": "number",
            "description": "Associated company ID",
            "example": 1
          },
          "contactIds": {
            "description": "Array of associated contact IDs",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      }
    }
  }
}